Rename applicationId and package name
